Create Logistic regression with Mnist Dataset
This project was done with python using the pytorch library. We first import the Mnist dataset from Pytorch itself, we then convert it to a tensor data format and splitting it between validation and test sets. We then load said data into the data loader to implement some data augmentations such as batch size and shuffling of data. We then create a function called "fit" to store the problem independent parts such as learning rate, number of epochs and the type of optimization function. The last part is the creation of the model itself where we will house the "problem specific parts". We created a class called the "MnistModel" with a super/parent class called "nn.module". Here we include the __init__ function which is used to create the layers of the model, the forward function is used to flatten/reshape the data itself before returning it to the __init__ function. The other functions listed below are functions being defined for the use in the "fit" function.